feat: history pushes, redirects & minor refactors/fixes
