+++ server/src/main.rs
@@ -60,7 +60,7 @@ async fn main() {
     tracing_subscriber::registry()
         .with(
             tracing_subscriber::EnvFilter::try_from_default_env()
-                .unwrap_or_else(|_| "example_websockets=debug,tower_http=info,diesel=debug".into()),
+                .unwrap_or_else(|_| "example_websockets=debug,tower_http=debug                                      ,diesel=debug".into()),
         )
         .with(tracing_subscriber::fmt::layer())
         .init();
