+++ web/src/routes/(portal)/app/contacts/+page.server.js
@@ -1,0 +1,6 @@
+export function load({ url }) {
+    const contactId = url.searchParams.get('c');
+    return {
+        requestedContactId: contactId
+    };
+}
