+++ web/src/routes/(portal)/app/contacts/[contact_id]/+page.server.js
@@ -1,0 +1,7 @@
+import {redirect} from "@sveltejs/kit";
+
+export function load({ params }) {
+    const contactId = params.contact_id;
+
+    throw redirect(303, `/app/contacts?c=${contactId}`);
+}
